Centralize fixture uses of setup_n_nodes
