+++ tests/conftest.py
@@ -5,6 +5,8 @@ import pytest
 import pytest_asyncio
 import tempfile
 
+from tests.setup_nodes import setup_node_and_wallet, setup_n_nodes
+
 # Set spawn after stdlib imports, but before other imports
 multiprocessing.set_start_method("spawn")
 
@@ -141,3 +143,27 @@ if os.getenv("_PYTEST_RAISE", "0") != "0":
     @pytest.hookimpl(tryfirst=True)
     def pytest_internalerror(excinfo):
         raise excinfo.value
+
+
+@pytest_asyncio.fixture(scope="function")
+async def wallet_node(self_hostname):
+    async for _ in setup_node_and_wallet(test_constants, self_hostname):
+        yield _
+
+
+@pytest_asyncio.fixture(scope="function")
+async def three_nodes(db_version, self_hostname):
+    async for _ in setup_n_nodes(test_constants, 3, db_version=db_version, self_hostname=self_hostname):
+        yield _
+
+
+@pytest_asyncio.fixture(scope="function")
+async def four_nodes(db_version, self_hostname):
+    async for _ in setup_n_nodes(test_constants, 4, db_version=db_version, self_hostname=self_hostname):
+        yield _
+
+
+@pytest_asyncio.fixture(scope="function")
+async def five_nodes(db_version, self_hostname):
+    async for _ in setup_n_nodes(test_constants, 5, db_version=db_version, self_hostname=self_hostname):
+        yield _
